Fixed some errors, Smart Bubble Sort for stack not working yet

+++ src/com/os/InsertionSorter.java
@@ -43,42 +43,42 @@ public class InsertionSorter {
 
         int size = stack.size();
 
-        for (int i=0; i<size;i++)
-        {
+        for (int i = 0; i < size; i++) {
             Iterator<Pokemon> iterator = stack.iterator();
-            Pokemon min = iterator.next();
+                if(iterator.hasNext()) {
+                Pokemon min = iterator.next();
 
-            for (int j = 0; j < stack.size()-1; j++) {  //looking for the minimum
-                Pokemon current = iterator.next();
+                for (int j = 0; j < stack.size() - 1; j++) {  //looking for the minimum
+                    Pokemon current = iterator.next();
 
-                if (min.compareTo(current) < 0 ) {
+                    if (min.compareTo(current) < 0) {
 
-                    min = current;
+                        min = current;
+
+                    }
 
                 }
+                //min is found
 
-            }
-            //min is found
-
-            //copy of the original queue
-            Stack<Pokemon> tstack = new Stack<>();
-            for(Pokemon s: stack){
-                tstack.push(s);
-            }
-            //clear of the original queue
-            stack.deleteAll();
-
-            //search of the min element by inserting everything that is not the min element back to the queue
-            for (Pokemon s : tstack)
-            {
-                if(s.compareTo(min)==0){
-                    //the elements we're searching for
-                }else{
-                    stack.push(s);
+                //copy of the original queue
+                Stack<Pokemon> tstack = new Stack<>();
+                for (Pokemon s : stack) {
+                    tstack.push(s);
                 }
-            }
+                //clear of the original queue
+                stack.deleteAll();
 
-            sorted.push(min);
+                //search of the min element by inserting everything that is not the min element back to the queue
+                for (Pokemon s : tstack) {
+                    if (s.compareTo(min) == 0) {
+                        //the elements we're searching for
+                    } else {
+                        stack.push(s);
+                    }
+                }
+
+                sorted.push(min);
+            }
         }
         return sorted;
     }
@@ -89,42 +89,42 @@ public class InsertionSorter {
 
         int size = queue.size();
 
-        for (int i=0; i<size;i++)
-        {
+        for (int i=0; i<size;i++) {
             Iterator<Pokemon> iterator = queue.iterator();
-            Pokemon min = iterator.next();
+            if (iterator.hasNext()) {
+                Pokemon min = iterator.next();
 
-            for (int j = 0; j < queue.size()-1; j++) {  //looking for the minimum
-                Pokemon current = iterator.next();
+                for (int j = 0; j < queue.size() - 1; j++) {  //looking for the minimum
+                    Pokemon current = iterator.next();
 
-                if (min.compareTo(current) >= 1 ) {
+                    if (min.compareTo(current) >= 1) {
 
-                    min = current;
+                        min = current;
+
+                    }
 
                 }
+                //min is found
 
-            }
-            //min is found
-
-            //copy of the original queue
-            MyQueue<Pokemon> queue1 = new MyQueue<>();
-            for(Pokemon q: queue){
-                queue1.insert(q);
-            }
-            //clear of the original queue
-            queue.deleteAll();
-
-            //search of the min element by inserting everything that is not the min element back to the queue
-            for (Pokemon q : queue1)
-            {
-                if(q.compareTo(min)==0){
-                    //the elements we're searching for
-                }else{
-                    queue.insert(q);
+                //copy of the original queue
+                MyQueue<Pokemon> queue1 = new MyQueue<>();
+                for (Pokemon q : queue) {
+                    queue1.insert(q);
                 }
-            }
+                //clear of the original queue
+                queue.deleteAll();
 
-            sorted.insert(min);
+                //search of the min element by inserting everything that is not the min element back to the queue
+                for (Pokemon q : queue1) {
+                    if (q.compareTo(min) == 0) {
+                        //the elements we're searching for
+                    } else {
+                        queue.insert(q);
+                    }
+                }
+
+                sorted.insert(min);
+            }
         }
 
         return sorted;
