+++ src/com/os/SmartBubbleSorter.java
@@ -4,7 +4,7 @@ import java.util.Iterator;
 
 public class SmartBubbleSorter {
 
-    //TODO
+
     public MyLinkedList<Pokemon> smartBubbleSort(MyLinkedList<Pokemon> list) {
 
         int listSize = list.size();
@@ -24,7 +24,7 @@ public class SmartBubbleSorter {
         return list;
     }
 
-    //TODO
+
     public MyDoublyLinkedList<Pokemon> smartBubbleSort(MyDoublyLinkedList<Pokemon> myLinkedList)
     {
         int listSize = myLinkedList.size();
@@ -53,50 +53,48 @@ public class SmartBubbleSorter {
 
         int size = stack.size();
 
-        for (int i=0; i<size-1;i++)
-        {
+        for (int i=0; i<size-1;i++) {
             Iterator<Pokemon> iterator = stack.iterator();
-            Pokemon min = iterator.next();
+
             boolean isSwaped = false;
-            for (int j = 0; j < stack.size()-1-i; j++) {  //looking for the minimum
-                Pokemon current = iterator.next();
+                Pokemon min = iterator.next();
+                for (int j = 0; j < stack.size() - 1 - i; j++) {  //looking for the minimum
+                    Pokemon current = iterator.next();
 
-                if (min.compareTo(current) < 0 ) {
+                    if (min.compareTo(current) < 0) {
 
-                    min = current;
+                        min = current;
 
-                    isSwaped = true;
+                        isSwaped = true;
+
+                    }
 
                 }
-
-            }
-            //min is found
+                //min is found
 
 
-            //copy of the original stack
-            Stack<Pokemon> stack1 = new Stack<>();
-            for(Pokemon q: stack){
-                stack1.push(q);
-            }
-            //clear of the original stack
-            stack.deleteAll();
-
-            //search of the min element by inserting everything that is not the min element back to the stack
-            for (Pokemon q : stack1)
-            {
-                if(q.compareTo(min)==0) {
-                    //the elements we're searching for
+                //copy of the original stack
+                Stack<Pokemon> stack1 = new Stack<>();
+                for (Pokemon q : stack) {
+                    stack1.push(q);
                 }
-                else{
-                    stack.push(q);
+                //clear of the original stack
+                stack.deleteAll();
+
+                //search of the min element by inserting everything that is not the min element back to the stack
+                for (Pokemon q : stack1) {
+                    if (q.compareTo(min) == 0) {
+                        //the elements we're searching for
+                    } else {
+                        stack.push(q);
+                    }
+                }
+
+                sorted.push(min);
+                if (!isSwaped) {
+                    break;
                 }
             }
-
-            sorted.push(min);
-            if (!isSwaped){
-                break;
-            }
-        }
 
         return sorted;
     }
@@ -109,42 +107,42 @@ public class SmartBubbleSorter {
 
         int size = queue.size();
 
-        for (int i=0; i<size;i++)
-        {
+        for (int i=0; i<size;i++) {
             Iterator<Pokemon> iterator = queue.iterator();
-            Pokemon min = iterator.next();
+            if (iterator.hasNext()) {
+                Pokemon min = iterator.next();
 
-            for (int j = 0; j < queue.size()-1; j++) {  //looking for the minimum
-                Pokemon current = iterator.next();
+                for (int j = 0; j < queue.size() - 1; j++) {  //looking for the minimum
+                    Pokemon current = iterator.next();
 
-                if (min.compareTo(current) >= 1 ) {
+                    if (min.compareTo(current) >= 1) {
 
-                    min = current;
+                        min = current;
+
+                    }
 
                 }
+                //min is found
 
-            }
-            //min is found
-
-            //copy of the original queue
-            MyQueue<Pokemon> queue1 = new MyQueue<>();
-            for(Pokemon q: queue){
-                queue1.insert(q);
-            }
-            //clear of the original queue
-            queue.deleteAll();
-
-            //search of the min element by inserting everything that is not the min element back to the queue
-            for (Pokemon q : queue1)
-            {
-                if(q.compareTo(min)==0){
-                    //the elements we're searching for
-                }else{
-                    queue.insert(q);
+                //copy of the original queue
+                MyQueue<Pokemon> queue1 = new MyQueue<>();
+                for (Pokemon q : queue) {
+                    queue1.insert(q);
                 }
-            }
+                //clear of the original queue
+                queue.deleteAll();
 
-            sorted.insert(min);
+                //search of the min element by inserting everything that is not the min element back to the queue
+                for (Pokemon q : queue1) {
+                    if (q.compareTo(min) == 0) {
+                        //the elements we're searching for
+                    } else {
+                        queue.insert(q);
+                    }
+                }
+
+                sorted.insert(min);
+            }
         }
 
         return sorted;
